Edit margin and padding values for a card
The purpose of this exercise was to edit the margin and padding around
three boxes and a button all inside a bigger box. The main struggle was
figuring out even spacing between the different interior boxes so that
they would align, but the 'auto' setting was very useful for this
purpose.

I also used the two numbers trick [ex: "padding: 8px 24px"] that would
set different values for the top and bottom [8px] and left and right
[24px]. Some other examples of this include:

 - "padding: 8px" = top, bottom, left, right are 8px
 - "padding: 8px 24px 4px 16px" = top, left, bottom, right are all different

Resolves: 02-margin-and-padding-2